feat: Handle buzzer mute in buzzer.cpp
We can reuse the same timer by defining a mute pattern, and with a
couple of booleans we can enforce that the mute can be set only if an
alarm is ringing, and also to resume the previous alarm after the mute
period.

However, this mute does not handle the priority levels: calling one of
the Buzzer_*_Start *will* stop the mute period and run the corresponding
pattern.